socket/final/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
python-dotenv>=1.0.0
websockets>=11.0.0
